+++ packages/ui/eslint.config.js
@@ -3,7 +3,6 @@ import js from '@eslint/js';
 import globals from 'globals';
 import tseslint from 'typescript-eslint';
 import { defineConfig, globalIgnores } from 'eslint/config';
-import { qwikEslint9Plugin } from 'eslint-plugin-qwik';
 
 const ignores = [
   "**/*.log",
@@ -51,7 +50,6 @@ export default defineConfig(
   globalIgnores(ignores),
   js.configs.recommended,
   tseslint.configs.recommendedTypeChecked,
-  qwikEslint9Plugin.configs.recommended,
   {
     files: ['src/**/*.{js,mjs,cjs,jsx,mjsx,ts,tsx,mtsx}'],
     languageOptions: {
