+++ packages/ui/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@birdflop/rgb",
+  "version": "0.1.0",
+  "description": "RGB gradient library for generating Minecraft text gradients.",
+  "main": "./lib/index.js",
+  "types": "./lib-types/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./lib/index.js",
+      "require": "./lib/index.js",
+      "types": "./lib-types/index.d.ts"
+    },
+    "./components/*": {
+      "import": "./lib/components/*",
+      "require": "./lib/components/*",
+      "types": "./lib-types/components/*"
+    }
+  },
+  "files": [
+    "lib",
+    "lib-types"
+  ],
+  "engines": {
+    "node": ">=15.0.0"
+  },
+  "license": "AGPL-3.0-or-later",
+  "private": false,
+  "type": "module",
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build": "tsc",
+    "dev": "echo \"No dev script specified\" && exit 0",
+    "lint": "eslint \"src/**/*.ts*\"",
+    "test": "echo \"No test specified\" && exit 0"
+  },
+  "devDependencies": {
+    "@types/eslint": "^8.56.11",
+    "@types/node": "^22.5.0",
+    "@typescript-eslint/eslint-plugin": "^8.2.0",
+    "@typescript-eslint/parser": "^8.2.0",
+    "eslint": "^8.57.0",
+    "typescript": "5.5.4"
+  }
+}
